Rename locale.{c,h} to iperf_locale.{c,h} to avoid filename collision with system header <locale.h>.
This apparently fixes problems on an ARM build, but this was generally
broken anyway.  It's slightly amazing this didn't cause problems before;
perhaps we never used <locale.h> before?

Addresses #203.

(cherry picked from commit a1861d5f87476213d80228a2277941baea8e1156)

Conflicts:
	src/iperf_api.c
