+++ src/iperf_api.c
@@ -63,7 +63,7 @@
 #include "units.h"
 #include "tcp_window_size.h"
 #include "iperf_util.h"
-#include "locale.h"
+#include "iperf_locale.h"
 
 
 /* Forwards. */
